Clamp race wash colors before narrowing

+++ src/main/PAL/main/race/series_cleared_wash.c
@@ -3,7 +3,9 @@
 #include "game/render.h"
 #include "game/render_internal.h"
 
-static s32 ClampColorChannel(s32 value) {
+#include <stdint.h>
+
+static s32 ClampColorChannel(int64_t value) {
     if (value < 0) {
         return 0;
     }
@@ -13,8 +15,10 @@ static s32 ClampColorChannel(s32 value) {
 void DrawSeriesClearedWash(s32 washProgress, s32 fadeLevel) {
     GameOrderingTableEntry *ot = GamePrimaryOrderingTable(0);
     s32 red = ClampColorChannel(fadeLevel);
-    s32 green = ClampColorChannel(fadeLevel + washProgress / 8);
-    s32 blue = ClampColorChannel(fadeLevel + washProgress / 4);
+    s32 green = ClampColorChannel(
+        (int64_t)fadeLevel + washProgress / 8);
+    s32 blue = ClampColorChannel(
+        (int64_t)fadeLevel + washProgress / 4);
     u8 *next = GameQueueTileTrans(
         ot, RENDER_PRIM_CURSOR_AS(u8), 0, 0, 0x140, 0xF0,
         red, green, blue);
